+++ src-tauri/src/lib.rs
@@ -5,14 +5,20 @@
 mod clamav;
 mod sysinfo;
 
-use crate::clamav::scan::{start_custom_scan, start_full_scan, start_main_scan, stop_scan};
-use crate::clamav::update::{get_clamav_version, update_definitions};
-use crate::sysinfo::{get_sys_info, get_sys_stats};
-
 use specta::specta;
 use std::process::Command;
 use tauri::command;
 use tauri_specta::{collect_commands, Builder};
+
+// Local Crates
+use crate::clamav::quarantine::{
+    delete_quarantine, list_quarantine, quarantine_file, restore_quarantine,
+};
+use crate::clamav::scan::{start_custom_scan, start_full_scan, start_main_scan, stop_scan};
+use crate::clamav::update::{get_clamav_version, update_definitions};
+use crate::clamav::remove_file;
+use crate::clamav::bulk_actions::{quarantine_all,delete_all,restore_all,clear_quarantine};
+use crate::sysinfo::{get_sys_info, get_sys_stats};
 
 #[command]
 #[specta]
@@ -36,14 +42,22 @@ pub fn run() {
         start_custom_scan,
         stop_scan,
         update_definitions,
-        get_clamav_version
+        get_clamav_version,
+        quarantine_file,
+        list_quarantine,
+        restore_quarantine,
+        delete_quarantine,
+        remove_file,
+        quarantine_all,
+        delete_all,
+        restore_all,
+        clear_quarantine
     ]);
 
     tauri::Builder::default()
-        .plugin(tauri_plugin_fs::init())
+        .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_os::init())
-        .plugin(tauri_plugin_opener::init())
         .invoke_handler(builder.invoke_handler())
         .run(tauri::generate_context!())
         .expect("error while running tauri application");