+++ src/lib/constants/columns.tsx
@@ -1,60 +1,8 @@
 import { Button } from "@/components/ui/button";
-import { ArrowDown, ArrowUp, ArrowUpDown, BugOff, EyeOff, FolderOpen, Logs, MoreHorizontal, RotateCcw, ScrollText, Search, ShieldBan, Trash, Trash2 } from "lucide-react";
+import { ArrowDown, ArrowUp, ArrowUpDown, EyeOff, Logs, MoreHorizontal, ScrollText, Search, Trash2 } from "lucide-react";
 import { ColumnDef } from "@tanstack/react-table";
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator, DropdownMenuTrigger } from "@/components/ui/dropdown-menu";
-import { IHistoryData, IQuarantineData, ISchedulerData } from "@/lib/types";
-
-export const THREATS_COLS: ColumnDef<IQuarantineData>[] = [
-  {
-     accessorKey: "displayName",
-     header: ({column}) => (
-          <div className="flex items-center justify-between gap-2">
-               <span>Threat</span>
-               <Button variant="ghost" onClick={()=>column.toggleSorting(column.getIsSorted() === "asc")} size="icon-sm">
-                    {column.getIsSorted()==="asc" ? <ArrowUp className="h-4 w-4" /> : column.getIsSorted()==="desc" ? <ArrowDown className="h-4 w-4" /> : <ArrowUpDown className="h-4 w-4" />}
-               </Button>
-          </div>
-     )
-  },
-  {
-     accessorKey: "filePath",
-     header: "Path"
-  },
-  {
-     accessorKey: "detectedAt",
-     header: "Detected At"
-  },
-  {
-     accessorKey: "status",
-     header: "Status"
-  },
-  {
-     id: "actions",
-     cell: () => {
-          // const threat = row.original { row }
-     
-          return (
-               <DropdownMenu>
-                    <DropdownMenuTrigger asChild>
-                         <Button variant="ghost" className="h-8 w-8 p-0">
-                              <span className="sr-only">Open menu</span>
-                              <MoreHorizontal className="h-4 w-4" />
-                         </Button>
-                    </DropdownMenuTrigger>
-                    <DropdownMenuContent align="end">
-                         <DropdownMenuLabel>Actions</DropdownMenuLabel>
-                         <DropdownMenuSeparator/>
-                         <DropdownMenuItem><BugOff/> Quarantine</DropdownMenuItem>
-                         <DropdownMenuItem><EyeOff/> Ignore</DropdownMenuItem>
-                         <DropdownMenuItem><ShieldBan/> Block</DropdownMenuItem>
-                         <DropdownMenuItem className="text-destructive"><Trash className="text-destructive"/> Delete permanently</DropdownMenuItem>
-                         <DropdownMenuItem><FolderOpen/> Open Containing Folder</DropdownMenuItem>
-                    </DropdownMenuContent>
-               </DropdownMenu>
-          )
-     },
-  }
-]
+import { IHistoryData, ISchedulerData } from "@/lib/types";
 
 export const HISTORY_COLS: ColumnDef<IHistoryData>[] = [
      {
@@ -85,8 +33,9 @@ export const HISTORY_COLS: ColumnDef<IHistoryData>[] = [
      },
      {
           id: "actions",
-          cell: () => {
-               // const item = row.original // {row}
+          cell: ({row}) => {
+               const item = row.original
+               console.log(item)
                return (
                     <DropdownMenu>
                          <DropdownMenuTrigger asChild>
@@ -105,69 +54,6 @@ export const HISTORY_COLS: ColumnDef<IHistoryData>[] = [
                )
           },
      }
-]
-
-export const QUARANTINE_COLS: ColumnDef<IQuarantineData>[] = [
-  {
-     accessorKey: "displayName",
-     header: ({column}) => (
-          <div className="flex items-center justify-between gap-2">
-               <span>Threat</span>
-               <Button variant="ghost" onClick={()=>column.toggleSorting(column.getIsSorted() === "asc")} size="icon-sm">
-                    {column.getIsSorted()==="asc" ? <ArrowUp className="h-4 w-4" /> : column.getIsSorted()==="desc" ? <ArrowDown className="h-4 w-4" /> : <ArrowUpDown className="h-4 w-4" />}
-               </Button>
-          </div>
-     )
-  },
-  {
-     accessorKey: "filePath",
-     header: "Path"
-  },
-  {
-    accessorKey: "status",
-    header: ({column}) => (
-          <div className="flex items-center justify-between gap-2">
-               <span>Status</span>
-               <Button variant="ghost" onClick={()=>column.toggleSorting(column.getIsSorted() === "asc")} size="icon-sm">
-                    {column.getIsSorted()==="asc" ? <ArrowUp className="h-4 w-4" /> : column.getIsSorted()==="desc" ? <ArrowDown className="h-4 w-4" /> : <ArrowUpDown className="h-4 w-4" />}
-               </Button>
-          </div>
-     )
-  },
-  {
-     accessorKey: "detectedAt",
-     header: ({column}) => (
-          <div className="flex items-center justify-between gap-2">
-               <span>Detected At</span>
-               <Button variant="ghost" onClick={()=>column.toggleSorting(column.getIsSorted() === "asc")} size="icon-sm">
-                    {column.getIsSorted()==="asc" ? <ArrowUp className="h-4 w-4" /> : column.getIsSorted()==="desc" ? <ArrowDown className="h-4 w-4" /> : <ArrowUpDown className="h-4 w-4" />}
-               </Button>
-          </div>
-     )
-  },
-  {
-     id: "actions",
-     cell: () => {
-          // const threat = row.original { row }
-          return (
-               <DropdownMenu>
-                    <DropdownMenuTrigger asChild>
-                         <Button variant="ghost" className="h-8 w-8 p-0">
-                              <span className="sr-only">Open menu</span>
-                              <MoreHorizontal className="h-4 w-4" />
-                         </Button>
-                    </DropdownMenuTrigger>
-                    <DropdownMenuContent align="end">
-                         <DropdownMenuLabel>Actions</DropdownMenuLabel>
-                         <DropdownMenuSeparator/>
-                         <DropdownMenuItem><RotateCcw/> Restore</DropdownMenuItem>
-                         <DropdownMenuItem className="text-destructive"><Trash className="text-destructive"/> Delete permanently</DropdownMenuItem>
-                         <DropdownMenuItem><FolderOpen/> Open Containing Folder</DropdownMenuItem>
-                    </DropdownMenuContent>
-               </DropdownMenu>
-          )
-     },
-  }
 ]
 
 export const SCHEDULER_COLS: ColumnDef<ISchedulerData>[] = [
@@ -207,8 +93,9 @@ export const SCHEDULER_COLS: ColumnDef<ISchedulerData>[] = [
      },
      {
           id: "actions",
-          cell: () => {
-               // const item = row.original // {row}
+          cell: ({row}) => {
+               const item = row.original
+               console.log(item)
                return (
                     <DropdownMenu>
                          <DropdownMenuTrigger asChild>